requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Sock==0.7.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
pyserial==3.5
simple-websocket==0.10.1
python-dotenv==1.0.0
psutil==5.9.5
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
blinker==1.6.2
bcrypt==4.0.1
redis==5.0.1
pyOpenSSL==23.2.0
cryptography==41.0.4
email-validator==2.0.0
Flask-Limiter==3.5.0
Flask-SeaSurf==1.1.1
Flask-Talisman==1.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
eventlet==0.33.3
gevent==23.9.1
schedule==1.2.1
alembic==1.12.1
supervisor==4.2.5
prometheus-client==0.19.0
sentry-sdk==1.32.0
certbot==2.7.4
uwsgi==2.0.23
newrelic==9.1.0
Flask-Caching==2.0.2
Flask-Session==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
